requirements.txt
asgiref==3.5.2
blinker==1.4
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.1.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.9
ldap3==2.9.1
MarkupSafe==2.1.1
mitmproxy~=9.0.1
msgpack==1.0.4
passlib==1.7.4
protobuf==4.21.4
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dotenv~=0.21.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sortedcontainers==2.4.0
tornado==6.2
urwid==2.1.2
Werkzeug==2.2.1
wsproto==1.1.0
zstandard==0.18.0
async-generator==1.10
attrs==22.1.0
cffi==1.15.1
charset-normalizer==2.1.1
exceptiongroup==1.0.4
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
lxml==4.9.1
mysqlclient==1.4.0
outcome==1.2.0
packaging==21.3
pycountry==22.3.5
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0
zipp==3.10.0

selenium~=4.7.0
chompjs~=1.1.9